Another update to the emulation loop, and timer mechanism.
Timers are now internally represented as 16 bit registers that hold
values in milliseconds.  This makes updating them easier and
straightforward.

Also, timing for the ellapsed time between execution of operations
has changed.  Timing now starts AFTER the last instruction has been
executed for the current emulation loop, and ends right before the next
instruction in the following emulation loop.  This is important because
instructions that could potentially take a long time, like LD_Vx_K,
which hangs until a user presses a key, will cause the calculated
ellapsed time quite large, and will cause the emulator to execute a
large number of instrutions in succession.

TLDR: we are only measuring how long it takes the emulation loop to reach
the opcode execution time, and not including the time it takes to execute
said opcodes.

+++ src/chip8emu/chip8emu.h
@@ -168,11 +168,20 @@ struct _chip8 {
                 /* Memory address register */
                 uint16_t I;
                 /* 
-                 * The two timers below (DT and ST) are decremented at 60 Hz */
+                 * The two timers below (DT and ST) are decremented at 60 Hz 
+                 * Note: 
+                 * Chip-8 Programs will set these registers to values [0-255],
+                 * because they should be decremented at 60 Hz.  However, to 
+                 * make emulation most straightforward, when they are being
+                 * set, the values are converted to milliseconds instead, which
+                 * simply means they are multiplied by 16.67.  This also means
+                 * that they will have maximum value of 4250, so they must
+                 * be 16 bits wide instead of 8 on the Chip-8.
+                 */
                 /* Delay Timer register */
-                uint8_t DT;
+                uint16_t DT;
                 /* Sound Timer register */
-                uint8_t ST;
+                uint16_t ST;
                 /* Program counter */
                 uint16_t PC;
                 /* Stack Pointer */
